Data Iterator behavior abstraction.  New VDataIterator component.  Refactor VDataTable.  Add new tests and fix old ones.
