+++ src/components/VDataIterator/index.js
@@ -1,0 +1,7 @@
+import VDataIterator from './VDataIterator'
+
+VDataIterator.install = function install (Vue) {
+  Vue.component(VDataIterator.name, VDataIterator)
+}
+
+export default VDataIterator
