+++ src/components/VDataTable/VDataTable.spec.js
@@ -9,7 +9,9 @@ test('VDataTable.vue', () => {
       propsData: {
         pagination: {
           descending: false,
-          sortBy: 'col1'
+          sortBy: 'col1',
+          rowsPerPage: 1,
+          page: 1
         },
         headers: [
           { text: 'First Column', value: 'col1', class: 'a-string' },
@@ -52,9 +54,47 @@ test('VDataTable.vue', () => {
     expect('Application is missing <v-app> component.').toHaveBeenTipped()
   })
 
-  it('should match a snapshot', () => {
+  it('should match a snapshot - no matching results', () => {
     const data = dataTableTestData()
+    data.propsData.search = "asdf"
     const wrapper = mount(VDataTable, data)
+
+    expect(wrapper.html()).toMatchSnapshot()
+    expect('Application is missing <v-app> component.').toHaveBeenTipped()
+  })
+
+  it('should match a snapshot - no data', () => {
+    const data = dataTableTestData()
+    data.propsData.items = []
+    const wrapper = mount(VDataTable, data)
+
+    expect(wrapper.html()).toMatchSnapshot()
+    expect('Application is missing <v-app> component.').toHaveBeenTipped()
+  })
+
+  it('should match a snapshot - with data', () => {
+    const data = dataTableTestData()
+    data.propsData.pagination.rowsPerPage = 3
+
+    const vm = new Vue()
+    const items = props => vm.$createElement('td', [props.item.col2])
+    const component = Vue.component('test', {
+      components: {
+        VDataTable
+      },
+      render (h) {
+        return h('v-data-table', {
+          props: {
+            ...data.propsData
+          },
+          scopedSlots: {
+            items
+          }
+        })
+      }
+    })
+
+    const wrapper = mount(component)
 
     expect(wrapper.html()).toMatchSnapshot()
     expect('Application is missing <v-app> component.').toHaveBeenTipped()
